Make hot plug function optional
HotPlug is supported by QDeviceWatcher, and there are still
some compatibility issues about it. Add a define symbol to
make it's support optional.

Add HOT_PLUG to DEFINES in the .pro file to enable it, comment
it out to disable it.

MUST REBUILD THE PROJECT AFTER MAKING CHANGES TO THE DEFINES
in the .pro file.

Signed-off-by: Song Qiang <songqiang1304521@gmail.com>